move plotly related tests from test_suf_plotting to test_plotly_backend
