+++ nilearn/plotting/surface/tests/test_surf_plotting.py
@@ -3,14 +3,12 @@
 # ruff: noqa: ARG001
 
 import tempfile
-from unittest import mock
 
 import numpy as np
 import pandas as pd
 import pytest
-from matplotlib.figure import Figure
 
-from nilearn._utils.helpers import is_kaleido_installed, is_plotly_installed
+from nilearn._utils.helpers import is_plotly_installed
 from nilearn.datasets import fetch_surf_fsaverage
 from nilearn.plotting import (
     plot_img_on_surf,
@@ -20,13 +18,6 @@ from nilearn.plotting import (
     plot_surf_stat_map,
 )
 from nilearn.plotting.displays import PlotlySurfaceFigure, SurfaceFigure
-
-try:
-    import IPython.display  # noqa:F401
-except ImportError:
-    IPYTHON_INSTALLED = False
-else:
-    IPYTHON_INSTALLED = True
 
 
 @pytest.mark.parametrize(
@@ -78,6 +69,15 @@ def test_surface_figure():
     assert s.output_file == "bar.png"
 
 
+def test_surface_figure_add_contours_raises_not_implemented():
+    """Test that calling add_contours method of SurfaceFigure raises a \
+    NotImplementedError.
+    """
+    figure = SurfaceFigure()
+    with pytest.raises(NotImplementedError):
+        figure.add_contours()
+
+
 @pytest.mark.skipif(
     is_plotly_installed(),
     reason=("This test only runs if Plotly is not installed."),
@@ -90,244 +90,12 @@ def test_plotly_surface_figure_import_error():
         PlotlySurfaceFigure()
 
 
-@pytest.mark.skipif(
-    not is_plotly_installed() or is_kaleido_installed(),
-    reason=("This test only runs if Plotly is installed, but not kaleido."),
-)
-def test_plotly_surface_figure_savefig_error():
-    """Test that an ImportError is raised when saving \
-       a PlotlySurfaceFigure without having kaleido installed.
-    """
-    with pytest.raises(ImportError, match="`kaleido` is required"):
-        PlotlySurfaceFigure().savefig()
-
-
-@pytest.mark.skipif(
-    not is_plotly_installed() or not is_kaleido_installed(),
-    reason=("Plotly and/or kaleido not installed; required for this test."),
-)
-def test_plotly_surface_figure():
-    ps = PlotlySurfaceFigure()
-    assert ps.output_file is None
-    assert ps.figure is None
-    ps.show()
-    with pytest.raises(ValueError, match="You must provide an output file"):
-        ps.savefig()
-    ps.savefig("foo.png")
-
-
-@pytest.mark.skipif(
-    not is_plotly_installed()
-    or not IPYTHON_INSTALLED
-    or not is_kaleido_installed(),
-    reason=(
-        "Plotly, Kaleido and/or Ipython is not installed; required for this"
-        " test."
-    ),
-)
-@pytest.mark.parametrize("renderer", ["png", "jpeg", "svg"])
-def test_plotly_show(renderer):
-    import plotly.graph_objects as go
-
-    ps = PlotlySurfaceFigure(go.Figure())
-    assert ps.output_file is None
-    assert ps.figure is not None
-    with mock.patch("IPython.display.display") as mock_display:
-        ps.show(renderer=renderer)
-    assert len(mock_display.call_args.args) == 1
-    key = "svg+xml" if renderer == "svg" else renderer
-    assert f"image/{key}" in mock_display.call_args.args[0]
-
-
-@pytest.mark.skipif(
-    not is_plotly_installed() or not is_kaleido_installed(),
-    reason=("Plotly and/or kaleido not installed; required for this test."),
-)
-def test_plotly_savefig(tmp_path):
-    import plotly.graph_objects as go
-
-    ps = PlotlySurfaceFigure(go.Figure(), output_file=tmp_path / "foo.png")
-    assert ps.output_file == tmp_path / "foo.png"
-    assert ps.figure is not None
-    ps.savefig()
-    assert (tmp_path / "foo.png").exists()
-
-
-@pytest.mark.parametrize("input_obj", ["foo", Figure(), ["foo", "bar"]])
-def test_instantiation_error_plotly_surface_figure(plotly, input_obj):
-    with pytest.raises(
-        TypeError,
-        match=("`PlotlySurfaceFigure` accepts only plotly figure objects."),
-    ):
-        PlotlySurfaceFigure(input_obj)
-
-
-def test_value_error_get_faces_on_edge(plotly, in_memory_mesh):
-    """Test that calling _get_faces_on_edge raises a ValueError when \
-       called with with indices that do not form a region.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError, match=("Vertices in parcellation do not form region.")
-    ):
-        figure._get_faces_on_edge([91])
-
-
-def test_plot_surf_contours_errors_with_plotly_figure(plotly, in_memory_mesh):
-    """Test that plot_surf_contours rasises error when given plotly obj."""
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(ValueError):
-        plot_surf_contours(in_memory_mesh, np.ones((10,)), figure=figure)
-
-
-def test_plot_surf_contours_errors_with_plotly_axes(plotly, in_memory_mesh):
-    """Test that plot_surf_contours rasises error when given plotly \
-        obj as axis.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(ValueError):
-        plot_surf_contours(in_memory_mesh, np.ones((10,)), axes=figure)
-
-
-def test_plotly_surface_figure_warns_on_isolated_roi(plotly, in_memory_mesh):
-    """Test that a warning is generated for ROIs with isolated vertices."""
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    # the method raises an error because the (randomly generated)
-    # vertices don't form regions
-    try:
-        with pytest.raises(UserWarning, match="contains isolated vertices:"):
-            figure.add_contours(levels=[0], roi_map=np.array([0, 1] * 10))
-    except Exception:
-        pass
-
-
-def test_distant_line_segments_detected_as_not_intersecting(plotly):
-    """Test that distant lines are detected as not intersecting."""
-    assert not PlotlySurfaceFigure._do_segs_intersect(0, 0, 1, 1, 5, 5, 6, 6)
-
-
-@pytest.mark.parametrize("levels,labels", [([0], ["a", "b"]), ([0, 1], ["a"])])
-def test_value_error_add_contours_levels_labels(
-    plotly, levels, labels, in_memory_mesh
-):
-    """Test that add_contours raises a ValueError when called with levels and \
-    labels that have incompatible lengths.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError,
-        match=("levels and labels need to be either the same length or None."),
-    ):
-        figure.add_contours(
-            levels=levels, labels=labels, roi_map=np.ones((10,))
-        )
-
-
-@pytest.mark.parametrize(
-    "levels,lines",
-    [([0], [{}, {}]), ([0, 1], [{}, {}, {}])],
-)
-def test_value_error_add_contours_levels_lines(
-    plotly, levels, lines, in_memory_mesh
-):
-    """Test that add_contours raises a ValueError when called with levels and \
-    lines that have incompatible lengths.
-    """
-    figure = plot_surf(in_memory_mesh, engine="plotly")
-    with pytest.raises(
-        ValueError,
-        match=("levels and lines need to be either the same length or None."),
-    ):
-        figure.add_contours(levels=levels, lines=lines, roi_map=np.ones((10,)))
-
-
 @pytest.fixture
 def surf_roi_data(rng, in_memory_mesh):
     roi_map = np.zeros((in_memory_mesh.n_vertices, 1))
     roi_idx = rng.integers(0, in_memory_mesh.n_vertices, size=10)
     roi_map[roi_idx] = 1
     return roi_map
-
-
-def test_add_contours(plotly, surface_image_roi):
-    """Test that add_contours updates data in PlotlySurfaceFigure."""
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi)
-    assert len(figure.figure.to_dict().get("data")) == 4
-
-    figure.add_contours(surface_image_roi, levels=[1])
-    assert len(figure.figure.to_dict().get("data")) == 5
-
-
-@pytest.mark.parametrize("hemi", ["left", "right", "both"])
-def test_add_contours_hemi(
-    plotly,
-    surface_image_roi,
-    hemi,
-):
-    """Test that add_contours works with all hemi inputs."""
-    if hemi == "both":
-        n_vertices = surface_image_roi.mesh.n_vertices
-    else:
-        n_vertices = surface_image_roi.data.parts[hemi].shape[0]
-    figure = plot_surf(
-        surface_image_roi.mesh,
-        engine="plotly",
-        hemi=hemi,
-    )
-    figure.add_contours(surface_image_roi)
-    assert figure._coords.shape[0] == n_vertices
-
-
-def test_add_contours_plotly_surface_image(plotly, surface_image_roi):
-    """Test that add_contours works with SurfaceImage."""
-    figure = plot_surf(
-        surf_map=surface_image_roi, hemi="left", engine="plotly"
-    )
-    figure.add_contours(roi_map=surface_image_roi)
-
-
-def test_surface_figure_add_contours_raises_not_implemented(plotly):
-    """Test that calling add_contours method of SurfaceFigure raises a \
-    NotImplementedError.
-    """
-    figure = SurfaceFigure()
-    with pytest.raises(NotImplementedError):
-        figure.add_contours()
-
-
-def test_add_contours_has_name(plotly, surface_image_roi):
-    """Test that contours added to a PlotlySurfaceFigure can be named."""
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, levels=[1], labels=["x"])
-    assert figure.figure.to_dict().get("data")[2].get("name") == "x"
-
-
-def test_add_contours_lines_duplicated(plotly, surface_image_roi):
-    """Test that the specifications of length 1 line provided to \
-     add_contours are duplicated to all requested contours.
-    """
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, lines=[{"width": 10}])
-    newlines = figure.figure.to_dict().get("data")[2:]
-    assert all(x.get("line").__contains__("width") for x in newlines)
-
-
-@pytest.mark.parametrize(
-    "key,value",
-    [
-        ("color", "yellow"),
-        ("width", 10),
-    ],
-)
-def test_add_contours_line_properties(plotly, key, value, surface_image_roi):
-    """Test that the specifications of a line provided to add_contours are \
-    stored in the PlotlySurfaceFigure data.
-    """
-    figure = plot_surf(surface_image_roi.mesh, engine="plotly")
-    figure.add_contours(surface_image_roi, levels=[1], lines=[{key: value}])
-    newline = figure.figure.to_dict().get("data")[2].get("line")
-    assert newline.get(key) == value
 
 
 def test_plot_surf_engine_error(in_memory_mesh):
@@ -387,24 +155,6 @@ def test_plot_surf(
     )
 
 
-@pytest.mark.skipif(
-    not is_plotly_installed() or is_kaleido_installed(),
-    reason=("This test only runs if Plotly is installed, but not kaleido."),
-)
-def test_plot_surf_error_when_kaleido_missing(
-    tmp_path, in_memory_mesh, bg_map
-):
-    with pytest.raises(ImportError, match="Saving figures"):
-        engine = "plotly"
-        # Plot with non None output file
-        plot_surf(
-            in_memory_mesh,
-            bg_map=bg_map,
-            engine=engine,
-            output_file=tmp_path / "tmp.png",
-        )
-
-
 @pytest.mark.parametrize("view", ["anterior", "posterior"])
 @pytest.mark.parametrize("hemi", ["left", "right", "both"])
 def test_plot_surf_hemi_views_plotly(
@@ -453,21 +203,6 @@ def test_plot_surf_error(engine, rng, in_memory_mesh):
             in_memory_mesh,
             surf_map=rng.standard_normal(size=(in_memory_mesh.n_vertices, 2)),
             engine=engine,
-        )
-
-
-@pytest.mark.parametrize("kwargs", [{"avg_method": "mean"}, {"alpha": "auto"}])
-def test_plot_surf_warnings_not_implemented_in_plotly(
-    plotly, kwargs, in_memory_mesh, bg_map
-):
-    with pytest.warns(
-        UserWarning, match="is not implemented for the plotly engine"
-    ):
-        plot_surf(
-            in_memory_mesh,
-            surf_map=bg_map,
-            engine="plotly",
-            **kwargs,
         )
 
 
@@ -533,16 +268,6 @@ def test_plot_surf_stat_map_with_threshold(
         stat_map=bg_map,
         threshold=0.3,
         engine=engine,
-    )
-
-
-def test_plot_surf_stat_map_colorbar_tick(plotly, in_memory_mesh, bg_map):
-    """Change colorbar tick format."""
-    plot_surf_stat_map(
-        in_memory_mesh,
-        stat_map=bg_map,
-        cbar_tick_format="%.2g",
-        engine="plotly",
     )
 
 
@@ -616,23 +341,6 @@ def test_plot_surf_roi_cmap_as_lookup_table(surface_image_roi):
         )
 
 
-@pytest.mark.parametrize("colorbar", [True, False])
-@pytest.mark.parametrize("cbar_tick_format", ["auto", "%f"])
-def test_plot_surf_parcellation_plotly(
-    plotly,
-    colorbar,
-    surface_image_parcellation,
-    cbar_tick_format,
-):
-    plot_surf_roi(
-        surface_image_parcellation.mesh,
-        roi_map=surface_image_parcellation,
-        engine="plotly",
-        colorbar=colorbar,
-        cbar_tick_format=cbar_tick_format,
-    )
-
-
 @pytest.mark.parametrize("engine", ["matplotlib", "plotly"])
 def test_plot_surf_roi_error(engine, rng, in_memory_mesh, surf_roi_data):
     if not is_plotly_installed() and engine == "plotly":
@@ -669,38 +377,6 @@ def test_plot_surf_roi_error(engine, rng, in_memory_mesh, surf_roi_data):
         match="Non-integer values in roi_map will no longer be allowed",
     ):
         plot_surf_roi(in_memory_mesh, roi_map=surf_roi_data, engine=engine)
-
-
-@pytest.mark.skipif(
-    not is_plotly_installed(),
-    reason=("This test only runs if Plotly is installed."),
-)
-@pytest.mark.parametrize(
-    "kwargs", [{"vmin": 2}, {"vmin": 2, "threshold": 5}, {"threshold": 5}]
-)
-def test_plot_surf_roi_colorbar_vmin_equal_across_engines(
-    matplotlib_pyplot, kwargs, in_memory_mesh
-):
-    """See issue https://github.com/nilearn/nilearn/issues/3944."""
-    roi_map = np.arange(0, len(in_memory_mesh.coordinates))
-
-    mpl_plot = plot_surf_roi(
-        in_memory_mesh,
-        roi_map=roi_map,
-        colorbar=True,
-        engine="matplotlib",
-        **kwargs,
-    )
-    plotly_plot = plot_surf_roi(
-        in_memory_mesh,
-        roi_map=roi_map,
-        colorbar=True,
-        engine="plotly",
-        **kwargs,
-    )
-    assert (
-        mpl_plot.axes[-1].get_ylim()[0] == plotly_plot.figure.data[1]["cmin"]
-    )
 
 
 @pytest.mark.parametrize(
